Major 0.2.0 update.
Replace PyQt4 with PySide.
Add esky support with updater that checks for new versions using a proxy
aware internet connection.
Add utility to check for duplicate calculations.
Remove UTR measures for periods of time since they shouldn't be used on
their own.
Add Inno Setup script for generating an installer.
Use a try catch statement for file opening, which prints an error
message to the status bar if the selected file cannot be opened.
List extended measures that haven't been changed when the the Comply
with UTR utility is run.