Remove update functionality.
The esky application updater doesn't work on Windows 7 and above. A more
ideal updater can be created using Inno Setup and a version file hosted
on a server.

While the Inno updater won't be written for sometime, I believe it's
best to to remove the esky updater for now, so that deployment is
consistent across Windows versions and so that the thinX may be deployed
on systems running Python 3.3.